pyproject.toml
[project]
name = "pvz-wiki-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "lxml>=5.4.0",
    "requests>=2.32.3",
    "opencc-python-reimplemented>=0.1.7",
]
